rename kebab cased files
