+++ webview/src/util/numberFormatting.ts
